# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive file server API built with Go. It supports file and folder management, sharing, real-time updates via WebSockets, and more. All protected endpoints require a Bearer Token for authorization. The API implements rate limiting: most routes are limited to 60 requests/second per IP, with stricter limits on login endpoints.",
        "title": "File Server API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/nodes/{nodeId}/share": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Shares a file or folder with another user, granting them read or write permissions.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shares"
                ],
                "summary": "Share a node",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node ID to share",
                        "name": "nodeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Share details",
                        "name": "shareRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.ShareRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ShareResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid request or cannot share with oneself",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - Node or recipient not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict - Node is already shared with this user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/shares/incoming/nodes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists files and folders shared with the current user by a specific sharer. Can list the root of shared items (by omitting parent_id) or the content of a shared subfolder (by providing its parent_id).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shares"
                ],
                "summary": "List items shared by a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username of the person who shared the content",
                        "name": "sharer_username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the shared parent folder to list. Omit for the root of shared items.",
                        "name": "parent_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Number of items to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order. Comma-separated list of fields. Use '-' for descending. E.g., 'type,-name'",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/serwer-plikow_internal_models.RichNode"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Missing sharer_username",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - Sharer or folder not found, or access denied",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shares/incoming/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gets a unique list of users who have shared one or more items with the currently authenticated user. This is the root level for the \"Shared with me\" view.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shares"
                ],
                "summary": "List users who shared with me",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api.SharingUserResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shares/incoming/writeable-folders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of folders shared by a specific user to which you have write access. This is primarily used for building a folder tree in a \"Move to...\" or \"Copy to...\" dialog within a shared context. It allows navigation by providing a `parent_id`.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shares"
                ],
                "summary": "List writeable shared folders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username of the person who owns the shared content",
                        "name": "sharer_username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The ID of the parent shared folder to list. Omit for the root level of folders you have write-access to from this user.",
                        "name": "parent_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/serwer-plikow_internal_models.RichNode"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Missing 'sharer_username' parameter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden - You have read-only access to the parent folder",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - Sharer user not found, or parent folder not found/access denied",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shares/outgoing/nodes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gets a paginated and sortable list of unique nodes (files and folders) that the currently authenticated user has shared with others.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shares"
                ],
                "summary": "List nodes I have shared",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Number of items to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order. Comma-separated list of fields. Use '-' for descending. E.g., 'type,-name'",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/serwer-plikow_internal_models.RichNode"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shares/{shareId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revokes a share entry. Only the original sharer can do this.",
                "tags": [
                    "shares"
                ],
                "summary": "Revoke a share",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the share to delete",
                        "name": "shareId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request - Invalid share ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - Share not found or you are not the owner",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api.ShareRequest": {
            "type": "object",
            "properties": {
                "permissions": {
                    "type": "string",
                    "enum": [
                        "read",
                        "write"
                    ],
                    "example": "read"
                },
                "recipient_username": {
                    "type": "string",
                    "example": "user2"
                }
            }
        },
        "internal_api.ShareResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 42
                },
                "node_id": {
                    "type": "string",
                    "example": "_vx2a-43VqRT5wz_s9u4"
                },
                "permissions": {
                    "type": "string",
                    "example": "read"
                },
                "recipient_id": {
                    "type": "integer",
                    "example": 2
                },
                "shared_at": {
                    "type": "string",
                    "example": "2025-12-01T15:04:05Z"
                },
                "sharer_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_api.SharingUserResponse": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string",
                    "example": "Jan Kowalski"
                },
                "id": {
                    "type": "integer",
                    "example": 2
                },
                "username": {
                    "type": "string",
                    "example": "user2"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}